requirements.txt
python-decouple==3.4
selenium==3.141.0
urllib3==1.26.2
